Relocate ui resources
