+++ ui/resources/LICENSE.txt
