+++ ui/resources/expanded.png
